Initial GTK version of HypView

+++ gtk/hv_curs.c
@@ -1,7 +1,8 @@
 #include "hv_gtk.h"
 
-void HypGetCursorPosition(DOCUMENT *doc, short x, short y, TEXT_POS *pos)
+void HypGetCursorPosition(DOCUMENT *doc, int x, int y, TEXT_POS *pos)
 {
+	/* YYY */
 	UNUSED(doc);
 	UNUSED(x);
 	UNUSED(y);
